change something

+++ src/model/models.py
@@ -21,7 +21,7 @@ class GeneralModel(nn.Module):
         self.checkpoint = checkpoint
         self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
         self.tokenizer = AutoTokenizer.from_pretrained(checkpoint)
-        self.base_model = AutoModelForSeq2SeqLM.from_pretrained(checkpoint).to(self.device)
+        self.base_model = AutoModelForSeq2SeqLM.from_pretrained(checkpoint, torch_dtype = torch.float16).to(self.device)
         self.rank = rank
         self.scaler = GradScaler()  # Thêm scaler cho mixed precision
